consumer/requirements.txt
astroid==2.4.2
attrs==19.3.0
flake8==3.8.3
importlib-metadata==1.7.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
pg8000==1.15.3
pika==1.1.0
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.5.3
pyparsing==2.4.7
pytest==5.4.3
pytest-mock==3.1.1
scramp==1.2.0
six==1.15.0
SQLAlchemy==1.3.18
SQLAlchemy-Utils==0.36.7
testing.common.database==2.0.3
testing.postgresql==1.3.0
toml==0.10.1
typed-ast==1.4.1
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.1.0
